requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
cachetools==7.0.5
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.3.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
executing==2.1.0
faiss-cpu==1.13.2
fastapi==0.135.2
fastjsonschema==2.21.1
filelock==3.25.2
fqdn==1.5.1
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.14.0
hf-xet==1.4.2
httpcore==1.0.7
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.5.3
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
mpmath==1.3.0
narwhals==2.18.1
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==12.1.1
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.48
protobuf==6.33.6
psutil==6.1.1
pure_eval==0.2.3
pyarrow==23.0.1
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2024.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2026.3.32
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.3.3
rpds-py==0.22.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
Send2Trash==1.8.3
sentence-transformers==5.3.0
setuptools==75.7.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==1.0.0
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
torch==2.11.0
tornado==6.4.2
tqdm==4.67.3
traitlets==5.14.3
transformers==5.4.0
typer==0.24.1
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.42.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
